refactor(doc): split element and style registry
Mirror the xls module: a StyleRegistry owns the resolved character
TextStyles (index 0 = default) and the interned font names, while the
ElementRegistry stores only a per-element style index. CharacterStyles
shrinks to a runs-only CharacterRuns; the SPRM/font-table resolution
moves to doc_style.{hpp,cpp}.
