+++ src/odr/internal/oldms/text/doc_element_registry.hpp
@@ -2,10 +2,8 @@
 
 #include <odr/definitions.hpp>
 #include <odr/document_element.hpp>
-#include <odr/style.hpp>
 
-#include <deque>
-#include <map>
+#include <cstdint>
 #include <string>
 #include <unordered_map>
 #include <vector>
@@ -42,20 +40,15 @@ public:
 
   void append_child(ElementIdentifier parent_id, ElementIdentifier child_id);
 
-  /// Stores a font name and returns a pointer that stays valid for the
-  /// registry's lifetime (`TextStyle::font_name` is a `const char *`).
-  const char *intern_font_name(const std::string &name);
-
-  /// Character style of a span or paragraph element.
-  void set_element_style(ElementIdentifier id, TextStyle style);
-  [[nodiscard]] const TextStyle *element_style(ElementIdentifier id) const;
+  /// Character style of a span or paragraph element, as an index into the
+  /// document's `StyleRegistry` (0 is the default style).
+  void set_element_style_index(ElementIdentifier id, std::uint32_t index);
+  [[nodiscard]] std::uint32_t element_style_index(ElementIdentifier id) const;
 
 private:
   std::vector<Element> m_elements;
   std::unordered_map<ElementIdentifier, Text> m_texts;
-  std::unordered_map<ElementIdentifier, TextStyle> m_styles;
-  /// Deque: elements never move, so interned `c_str()`s stay valid.
-  std::deque<std::string> m_font_names;
+  std::unordered_map<ElementIdentifier, std::uint32_t> m_style_indices;
 
   void check_element_id(ElementIdentifier id) const;
   void check_text_id(ElementIdentifier id) const;
